+++ apps/web/components/task/chat/message-history-search.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useMemo, useRef, useState } from "react";
 import { createPortal } from "react-dom";
-import { cn } from "@/lib/utils";
 import { searchHistory, type SearchHit } from "./message-history";
+import { HitRow } from "./message-history-search-row";
 
 type MessageHistorySearchProps = {
   isOpen: boolean;
@@ -106,35 +106,6 @@ function computeStyle(anchorRect: DOMRect | null): React.CSSProperties | null {
     zIndex: 60,
     pointerEvents: "auto",
   };
-}
-
-type HighlightedRowProps = {
-  hit: SearchHit;
-  isSelected: boolean;
-  rowIndex: number;
-  onMouseEnter: (rowIndex: number) => void;
-  onClick: (historyIndex: number) => void;
-};
-
-function HitRow({ hit, isSelected, rowIndex, onMouseEnter, onClick }: HighlightedRowProps) {
-  const firstLine = hit.content.split("\n", 1)[0];
-  return (
-    <button
-      type="button"
-      data-hit-index={rowIndex}
-      data-testid="history-search-row"
-      onMouseEnter={() => onMouseEnter(rowIndex)}
-      onClick={() => onClick(hit.index)}
-      className={cn(
-        "flex w-full cursor-pointer select-none items-center gap-2 rounded-[6px] mx-1 px-2 py-1.5 text-xs text-left",
-        "hover:bg-muted/50",
-        isSelected && "bg-muted/50",
-      )}
-      style={{ width: "calc(100% - 8px)" }}
-    >
-      <span className="truncate flex-1">{firstLine || "(empty)"}</span>
-    </button>
-  );
 }
 
 export function MessageHistorySearch({
